refactor(api): split up Dagger modules by APIs
